requirements.txt
# System Build Tools
cmake>=3.28.0  # Required to build dlib

# Web Frameworks (Backend)
Django>=5.1.0
Flask>=3.0.0
Flask-Login>=0.6.0
django-cors-headers>=4.3.0  # Essential for React-Django integration

# Database
pymongo>=4.6.0  # MongoDB Driver
certifi>=2024.0.0  # SSL Certificate Verification (Crucial for Atlas)
dnspython>=2.6.0  # MongoDB SRV support

# Computer Vision & Image Processing
opencv-python>=4.9.0
numpy>=1.26.0
face_recognition>=1.3.0
dlib>=19.24.0  # Face detection core (Requires C++ tools if wheel not available)
Pillow>=10.2.0

# Object Detection (YOLOv8)
ultralytics>=8.0.0

# Utilities
python-dotenv>=1.0.0  # Loads .env variables
requests>=2.31.0

# ==========================================
# FRONTEND DEPENDENCIES (Node.js/npm)
# ==========================================
# NOTE: These are NOT Python packages. Do not install with pip.
# To install frontend dependencies, navigate to the 'frontend' folder and run:
# npm install
#
# Core Libraries:
# - react @ ^18.3.1
# - react-dom @ ^18.3.1
# - bootstrap @ ^5.3.3
#
# Build Tools & Dev Dependencies:
# - vite @ ^5.4.1
# - @vitejs/plugin-react @ ^4.3.1
# - eslint @ ^9.9.0
# - @types/react @ ^18.3.3
